Update folder structure to MVC
