+++ src/main/java/com/example/ticketReservationSystem/controller/FlightController.java
@@ -1,0 +1,4 @@
+package com.example.ticketReservationSystem.controller;
+
+public class FlightController {
+}
