--- src/main/java/com/example/ticketReservationSystem/flight/FlightRepository.java
@@ -1,4 +1,0 @@
-package com.example.ticketReservationSystem.flight;
-
-public interface FlightRepository {
-}
