--- src/main/java/com/example/ticketReservationSystem/flight/FlightService.java
@@ -1,4 +1,0 @@
-package com.example.ticketReservationSystem.flight;
-
-public class FlightService {
-}
