+++ src/main/java/com/example/ticketReservationSystem/model/UserType.java
@@ -1,0 +1,4 @@
+package com.example.ticketReservationSystem.model;
+
+public class UserType {
+}
