+++ src/main/java/com/example/ticketReservationSystem/repository/FlightRepository.java
@@ -1,0 +1,4 @@
+package com.example.ticketReservationSystem.repository;
+
+public interface FlightRepository {
+}
