+++ src/main/java/com/example/ticketReservationSystem/repository/TicketRepository.java
@@ -1,5 +1,6 @@
-package com.example.ticketReservationSystem.ticket;
+package com.example.ticketReservationSystem.repository;
 
+import com.example.ticketReservationSystem.model.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TicketRepository extends JpaRepository<Ticket, Long> {
