+++ src/main/java/com/example/ticketReservationSystem/repository/UserRepository.java
@@ -1,7 +1,7 @@
-package com.example.ticketReservationSystem.user;
+package com.example.ticketReservationSystem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.example.ticketReservationSystem.ticket.Ticket;
+import com.example.ticketReservationSystem.model.Ticket;
 
 public interface UserRepository extends JpaRepository<Ticket, Long> {
 }
