+++ src/main/java/com/example/ticketReservationSystem/service/FlightService.java
@@ -1,0 +1,4 @@
+package com.example.ticketReservationSystem.service;
+
+public class FlightService {
+}
