+++ src/main/java/com/example/ticketReservationSystem/service/TicketService.java
@@ -1,0 +1,4 @@
+package com.example.ticketReservationSystem.service;
+
+public class TicketService {
+}
