+++ src/main/java/com/example/ticketReservationSystem/service/UserService.java
@@ -1,0 +1,4 @@
+package com.example.ticketReservationSystem.service;
+
+public class UserService {
+}
