--- src/main/java/com/example/ticketReservationSystem/ticket/TicketController.java
@@ -1,4 +1,0 @@
-package com.example.ticketReservationSystem.ticket;
-
-public class TicketController {
-}
