--- src/main/java/com/example/ticketReservationSystem/user/UserController.java
@@ -1,4 +1,0 @@
-package com.example.ticketReservationSystem.user;
-
-public class UserController {
-}
