--- src/main/java/com/example/ticketReservationSystem/userType/UserType.java
@@ -1,4 +1,0 @@
-package com.example.ticketReservationSystem.userType;
-
-public class UserType {
-}
